+++ src/main/java/com/example/holaserver/Auth/SocialUserInfoDto.java
@@ -1,4 +1,4 @@
-package com.example.holaserver.User;
+package com.example.holaserver.Auth;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
