+++ src/main/java/com/example/holaserver/User/User.java
@@ -33,7 +33,7 @@ public class User extends BaseTimeEntity {
 
     private Timestamp removedAt;
 
-    @Builder(builderClassName = "OAuth2Register", builderMethodName = "oauth2Register")
+    @Builder
     public User(String name, String email, Type type, String oauthType, Byte rating, String imgPath, String oauthIdentity) {
         this.name = name;
         this.email = email;
