+++ src/main/java/com/example/holaserver/User/UserRepository.java
@@ -7,6 +7,5 @@ import org.springframework.stereotype.Repository;
 public interface UserRepository extends JpaRepository<User, Long> {
     User findByName(String name);
     Boolean existsByOauthIdentityAndOauthType(String oauthIdentity, String oauthType);
-
     User findByOauthIdentity(String oauthIdentity);
 }
