+++ src/main/resources/application.yml
@@ -45,7 +45,7 @@ spring:
               - account_email
             client-name: Kakao #꼭 있어야하는건지는 모르겠음
             authorization-grant-type: authorization_code
-            redirect-uri: http://localhost:3000/login/oauth/kakao
+            redirect-uri: http://localhost:3000/api/login/oauth/kakao
             client-authentication-method: POST
         provider:
           kakao:
